Move interrupt handlers from the base shell to the kernel.
