+++ src/async_kernel/shell/base.py
@@ -1,27 +1,21 @@
 from __future__ import annotations
 
 import contextlib
-import os
-import signal
-import sys
-import time
-from typing import TYPE_CHECKING, Any, ClassVar, Generic, Literal, Self
+from typing import TYPE_CHECKING, Any, Generic, Literal, Self
 
-from aiologic.lowlevel import enable_signal_safety
 from traitlets import traitlets
 from traitlets.config.configurable import LoggingConfigurable
 from typing_extensions import override
 
 import async_kernel
-from async_kernel import Caller, utils
-from async_kernel.common import Fixed, KernelInterrupt
+from async_kernel import utils
+from async_kernel.common import Fixed
 from async_kernel.interface import HasInterface
 from async_kernel.pending import PendingManager
-from async_kernel.typing import Channel, ExecuteContent, Job, T_interface_co
+from async_kernel.typing import ExecuteContent, Job, T_interface_co
 
 if TYPE_CHECKING:
     from collections.abc import Callable, Generator
-    from types import FrameType
 
     from async_kernel import Kernel
     from async_kernel.typing import Content
@@ -65,12 +59,8 @@ class BaseShell(HasInterface[T_interface_co], LoggingConfigurable, Generic[T_int
         lambda c: None if c["owner"].is_mainshell else c["owner"].pending_manager.id
     )
 
-    "A set for callbacks to register for calling when `interrupt` is called."
-
     _execution_count = 0
     _resetting = False
-    _interrupt_requested: bool | Literal["FORCE"] = False
-    _interrupts: ClassVar[set[Callable[[], object]]] = set()
     _user_ns: Fixed[Self, dict] = Fixed(dict)
     _stop_on_error_info: Fixed[Self, dict[Literal["time", "execution_count"], Any]] = Fixed(dict)
 
@@ -148,16 +138,6 @@ class BaseShell(HasInterface[T_interface_co], LoggingConfigurable, Generic[T_int
             restore()
 
     def apply_patches(self) -> Callable[[], None]:
-
-        with contextlib.suppress(ValueError, AttributeError):
-            import signal  # noqa: PLC0415
-
-            sig = signal.signal(signal.SIGINT, self._signal_handler)
-
-            def restore() -> None:
-                signal.signal(signal.SIGINT, sig)
-
-            return restore
         return lambda: None
 
     def initialize(self) -> None:
@@ -179,67 +159,6 @@ class BaseShell(HasInterface[T_interface_co], LoggingConfigurable, Generic[T_int
             return
         if not self.is_mainshell:
             self.kernel._subshell_stopped(self)  # pyright: ignore[reportPrivateUsage]
-
-    @enable_signal_safety
-    def _signal_handler(self, signum, frame: FrameType | None) -> None:
-        "Handle interrupt signals."
-
-        match self._interrupt_requested:
-            case "FORCE":
-                self._interrupt_requested = False
-                raise KernelInterrupt
-            case True:
-                if frame and frame.f_locals is self.kernel.shell.user_ns:
-                    self._interrupt_requested = False
-                    raise KernelInterrupt
-                self.last_interrupt_frame = frame
-
-                def clearlast_interrupt_frame():
-                    if self.last_interrupt_frame is frame:
-                        self.last_interrupt_frame = None
-
-                def re_raise():
-                    if self.last_interrupt_frame is frame:
-                        self._interrupt_now(force=True)
-
-                # Race to check if the main thread should be interrupted.
-                self.kernel.callers[Channel.shell].call_direct(clearlast_interrupt_frame)
-                self.kernel.callers[Channel.control].call_later(1, re_raise)
-            case False:
-                signal.default_int_handler(signum, frame)
-
-    def _interrupt_now(self, *, force=False) -> None:
-        """
-        Request an interrupt of the currently running shell thread.
-
-        If called from the main thread, sets the interrupt request flag and sends a SIGINT signal
-        to the current process. On Windows, uses `signal.raise_signal`; on other platforms, uses `os.kill`.
-        If `force` is True, sets the interrupt request flag to "FORCE".
-
-        Args:
-            force: If True, requests a forced interrupt. Defaults to False.
-        """
-        # Restricted this to when the shell is running in the main thread.
-        if self.parent.callers[Channel.shell].id == Caller.CALLER_MAIN_THREAD_ID:
-            self._interrupt_requested = "FORCE" if force else True
-            if sys.platform == "win32":
-                signal.raise_signal(signal.SIGINT)
-                time.sleep(0)
-            else:
-                os.kill(os.getpid(), signal.SIGINT)
-
-    def interrupt(self) -> None:
-        """
-        Perform a keyboard interrupt.
-        """
-        debugger = self.kernel.debugger
-        if (sys.platform != "emscripten") and (not debugger.enabled or not debugger.stopped_threads):
-            self._interrupt_now()
-        while self._interrupts:
-            try:
-                self._interrupts.pop()()
-            except Exception:
-                pass
 
     def reset(self, new_session=True, aggressive=False) -> None:
         "Reset the shell, cancelling all associated pending."
